underground markers

+++ src/main.ts
@@ -1,70 +1,73 @@
+import boxen from 'boxen';
 import chalk from 'chalk';
+import { Spinner } from 'cli-spinner';
 import decompress from 'decompress';
+import dotenv from 'dotenv';
 import fs from 'fs';
 import PDFMerger from 'pdf-merger-js';
 import { createPDF } from './create';
 import { setHeader, startTime } from './header';
 import './styles';
 import { convertKmlToGeoJson } from './toGeoJson';
+dotenv.config();
 
 export const photos: any[] = [],
-             subs: any[] = [];
+             polesInLeftColumn: any[] = [],
+             polesInRightColumn: any[] = [],
+             undergroundsInLeftColumn: any[] = [],
+             undergroundsInRightColumn: any[] = [];
 
 const photoWidth = 120,
       photoHeight = 150,
-      photoGap = 30,
-      imagesDir = `src/static/images`,
-      noPhotoPath = `${imagesDir}/no-photo-infinitel.png`;
+      photoGap = 30;
 
-const inputFileDir = `input`,
-      inputFile = fs.readdirSync(inputFileDir),
-      kmlMediaDir = `src/kml-cloud-media`,
-      pdfsToMergeDir = `src/pdfs-to-merge`;
+const inputFileDir = 'input',
+      kmlMediaDir = 'src/kml-media',
+      pdfsToMergeDir = 'src/pdfs-to-merge',
+      imagesDir = 'src/static/images',
+      input = fs.readdirSync(inputFileDir),
+      noPhoto = `${imagesDir}/no-photo-infinitel.png`;
 
 let   GEOJSON: any,
-      POLES_IN_PDF = 96,
-      polesAmount = 0,
+      polesInPdf = 96,
+      polesAmountRecursive = 0,
       polesAmountImmutable = 0,
       polesWithoutPhoto = 0,
+      undergroundMarkersAmount = 0,
       photosAmount = 0,
-      subsMarkerAmount = 0,
-      subsAmount = 0,
-      subsAmountImmutable = 0,
       ignoredMarkers = 0,
       renamedMediaFolder = false;
 
 const merger = new PDFMerger();
 
+const spinner = new Spinner(chalk.yellow('Processando'));
+      spinner.setSpinnerString('|/-\\');
+
 (async() => {
-  const filename = inputFile[0];
+  const filename = input[0];
 
   if (filename) checkInputFileExtension(filename)
-  else throw new Error(chalk.redBright(`Coloque um arquivo KMZ na pasta input`))
+  else throw new Error(chalk.redBright('Coloque um arquivo KMZ na pasta input'))
 })()
 
 async function checkInputFileExtension(filename: string) {
   const extension = filename.split(`.`).pop();
 
   if (extension === 'kmz') {
-    console.log(chalk.blueBright(`KMZ Detectado...`));
+    console.log(chalk.blueBright('KMZ Detectado'));
     console.log('');
     
     await setHeader();
 
+    spinner.start();
+
     extractKmlFileAndMediaFolder(filename);
-  } else {
-    throw new Error(chalk.redBright(`O arquivo de entrada não é um KMZ`))
-  }
+  } else throw new Error(chalk.redBright('O arquivo de entrada não é um KMZ'))
 }
 
 async function extractKmlFileAndMediaFolder(filename: string) {
-  if (fs.existsSync(kmlMediaDir)) {
-    fs.rmSync(kmlMediaDir, { recursive: true, force: true });
-  }
-
-  if (fs.existsSync(pdfsToMergeDir)) {
-    fs.rmSync(pdfsToMergeDir, { recursive: true, force: true });
-  }
+  if (fs.existsSync(kmlMediaDir)) fs.rmSync(kmlMediaDir, { recursive: true, force: true });
+  if (fs.existsSync(pdfsToMergeDir)) fs.rmSync(pdfsToMergeDir, { recursive: true, force: true });
 
   fs.mkdirSync(kmlMediaDir);
   fs.mkdirSync(pdfsToMergeDir);
@@ -79,10 +82,7 @@ async function extractKmlFileAndMediaFolder(filename: string) {
     `${kmlMediaDir}/${filename}.zip`
   )
 
-  await decompress(
-    `${kmlMediaDir}/${filename}.zip`,
-    kmlMediaDir
-  )
+  await decompress(`${kmlMediaDir}/${filename}.zip`, kmlMediaDir);
 
   fs.rmSync(`${kmlMediaDir}/${filename}.zip`);
 
@@ -92,15 +92,15 @@ async function extractKmlFileAndMediaFolder(filename: string) {
       const itemsInFolder = fs.readdirSync(kmlMediaDir);
         
       if (itemsInFolder.length > 1) {
-        const cloudMediaDir = `${kmlMediaDir}/${dirent.name}`,
-              cloudMediaFiles = fs.readdirSync(cloudMediaDir);
+        const mediaDir = `${kmlMediaDir}/${dirent.name}`,
+              mediaFiles = fs.readdirSync(mediaDir);
     
-        photosAmount = cloudMediaFiles.length;
+        photosAmount = mediaFiles.length;
     
-        for await (const file of cloudMediaFiles) {
+        for await (const file of mediaFiles) {
           fs.renameSync(
-            `${cloudMediaDir}/${file}`,
-            `${cloudMediaDir}/${file}.png`
+            `${mediaDir}/${file}`,
+            `${mediaDir}/${file}.png`
           )
         }
     
@@ -116,11 +116,13 @@ async function getGeoJsonFromKml() {
 
   await GEOJSON.features.forEach((marker: any) => {
     if (marker.properties.name) {
-      const pole = JSON.stringify(marker.properties.name).replace(/["​]/g, ''),
-            photos = marker.properties.com_exlyo_mapmarker_images_with_ext as string;
+      const markerName = JSON.stringify(marker.properties.name).replace(/["​]/g, ''),
+            markerPhotos = marker.properties.com_exlyo_mapmarker_images_with_ext as string;
 
-      if (photos && !renamedMediaFolder) { 
-        const photosObject = JSON.parse(photos) as Array<any>,
+      if (markerPhotos && !renamedMediaFolder) {
+        const photosObject = JSON.parse(markerPhotos) as Array<any>,
+
+        // get the media folder name before the slash
         mediaFolderName = photosObject[0].file_rel_path.match(/(\S+)\//)[1] as string;
         
         fs.readdirSync(kmlMediaDir, { withFileTypes: true })
@@ -133,25 +135,34 @@ async function getGeoJsonFromKml() {
           })
       }
 
-      if (pole) {;
-        if (pole.includes('.')) {
-          console.log(chalk.bgYellow.black(`O marcador ${pole} foi ignorado pois ainda não há suporte a casas decimais`));
-        }
-
-        if (Number(pole) && !pole.includes('.')) polesAmount++;
+      if (markerName.includes('.')) {
+        console.log(chalk.bgYellow.black(`O marcador ${markerName} foi ignorado pois ainda não há suporte a casas decimais`));
       }
 
-      polesAmountImmutable = polesAmount;
+      // only number = pole
+      if (Number(markerName) && !markerName.includes('.')) polesAmountRecursive++;
+
+      polesAmountImmutable = polesAmountRecursive;
+
+      // starts with sub = underground
+      if (markerName.startsWith('sub')) {
+        const photosInMarker = JSON.parse(marker.properties.com_exlyo_mapmarker_images_with_ext);
+        
+        undergroundMarkersAmount ++;
+        undergroundPhotosAmount += photosInMarker.length;
+      }
     }
   })
+
+  
 
   for (let index = 0; index <= polesAmountImmutable; index++) {
     GEOJSON.features.filter((marker: any) => {
       if (marker.properties.name) {
-        const markerIdentifier = JSON.stringify(marker.properties.name).replace(/["​]/g, '');
+        const markerName = JSON.stringify(marker.properties.name).replace(/["​]/g, '');
 
         // parseFloat to identify decimal places
-        const pole = parseFloat(markerIdentifier),
+        const pole = parseFloat(markerName),
               photos = marker.properties.com_exlyo_mapmarker_images_with_ext,
               coordinates = marker.geometry.coordinates as Array<number>;
 
@@ -167,7 +178,7 @@ async function getGeoJsonFromKml() {
           rightPhoto = undefined;
         }
         
-        if (!pole.toString().includes('.') && pole === index) { 
+        if (!pole.toString().includes('.') && pole === index) {
           setLeftAndRightPhotos(
             pole,
             coordinates,
@@ -188,37 +199,32 @@ function setLeftAndRightPhotos(
   leftPhoto: any,
   rightPhoto: any
 ) {
-  const photosDir = 'src/kml-cloud-media/';
-
   let leftPhotoPath = '',
       rightPhotoPath = '';
 
   if (leftPhoto && rightPhoto) {
-    leftPhotoPath = `${photosDir}${leftPhoto.file_rel_path}${leftPhoto.file_extension}`;
-    rightPhotoPath = `${photosDir}${rightPhoto.file_rel_path}${rightPhoto.file_extension}`;
+    leftPhotoPath = `${kmlMediaDir}/${leftPhoto.file_rel_path}${leftPhoto.file_extension}`;
+    rightPhotoPath = `${kmlMediaDir}/${rightPhoto.file_rel_path}${rightPhoto.file_extension}`;
   }
 
   if (leftPhoto && !rightPhoto) {
-    leftPhotoPath = `${photosDir}${leftPhoto.file_rel_path}${leftPhoto.file_extension}`;
-    rightPhotoPath = noPhotoPath;
+    leftPhotoPath = `${kmlMediaDir}/${leftPhoto.file_rel_path}${leftPhoto.file_extension}`;
+    rightPhotoPath = noPhoto;
   }
 
   if (!leftPhoto && rightPhoto) {
-    leftPhotoPath = noPhotoPath;
-    rightPhotoPath = `${photosDir}${rightPhoto.file_rel_path}${rightPhoto.file_extension}`;
+    leftPhotoPath = noPhoto;
+    rightPhotoPath = `${kmlMediaDir}/${rightPhoto.file_rel_path}${rightPhoto.file_extension}`;
   }
 
   if (!leftPhoto && !rightPhoto) {
     polesWithoutPhoto++;
-    leftPhotoPath = noPhotoPath;
-    rightPhotoPath = noPhotoPath;
+    leftPhotoPath = noPhoto;
+    rightPhotoPath = noPhoto;
   }
 
   splitColumns(pole, coordinates, leftPhotoPath, rightPhotoPath);
 }
-
-export let polesInLeftColumn: any[][] = [],
-           polesInRightColumn: any[][] = [];
 
 function splitColumns(
   pole: number,
@@ -237,26 +243,24 @@ function splitColumns(
   if (
     polesInLeftColumn.length - polesInRightColumn.length > 1 ||
     polesInRightColumn.length - polesInLeftColumn.length > 1
-  ) {
-    throw new Error(chalk.redBright('Há algum marcador intruso ou faltando!'));
-  }
+  ) throw new Error(chalk.redBright('Há algum marcador intruso ou faltando!'));
 
-  const POLES_IN_MEMORY = polesInLeftColumn.length + polesInRightColumn.length;
-  
+  const polesInMemory = polesInLeftColumn.length + polesInRightColumn.length;
+
   if (
-    POLES_IN_MEMORY === POLES_IN_PDF ||
-    (polesAmount < POLES_IN_PDF && POLES_IN_MEMORY === polesAmount)
+    polesInMemory === polesInPdf ||
+    (polesAmountRecursive < polesInPdf && polesInMemory === polesAmountRecursive)
   ) {
-    createColumns(polesInLeftColumn, polesInRightColumn);
+    createColumns();
   }
 }
 
-async function createColumns(polesInLeftColumn: any[][], polesInRightColumn: any[][]) {
+async function createColumns() {
   polesInLeftColumn.forEach(async (tableLeft, i) => {
     let tableRight = polesInRightColumn[i];
 
-    if (!tableLeft) { tableLeft = [null, [0, 0, 0], noPhotoPath, noPhotoPath] };
-    if (!tableRight) { tableRight = [null, [0, 0, 0], noPhotoPath, noPhotoPath] };
+    if (!tableLeft) tableLeft = [null, [0, 0, 0], noPhoto, noPhoto];
+    if (!tableRight) tableRight = [null, [0, 0, 0], noPhoto, noPhoto];
 
     let pageBreak;
     if (photos.length % 4 === 0 && photos.length !== 0) pageBreak = 'before';
@@ -341,19 +345,135 @@ async function createColumns(polesInLeftColumn: any[][], polesInRightColumn: any
 
     photos.push(leftAndRightColumnWithTablesAndPhotos);
     
-    if (photos.length === POLES_IN_PDF / 2) {
-      polesAmount -= POLES_IN_PDF;
+    if (photos.length === polesInPdf / 2) {
+      polesAmountRecursive -= polesInPdf;
       
       await createPDF();
     }
     
     // last pdf
     if (
-      polesAmount < POLES_IN_PDF &&
-      photos.length === Math.ceil(polesAmount / 2)
+      polesAmountRecursive < polesInPdf &&
+      photos.length === Math.ceil(polesAmountRecursive / 2)
     ) {
       createPDF();
-      setTimeout(() => { mergePDFS() }, 1000)
+      handleUndergrounds();
+      // setTimeout(() => { mergePDFS() }, 1000)
+    }
+  })
+}
+
+let undergroundPhotosAmount = 0;
+
+function handleUndergrounds() {
+  for (let index = 0; index <= undergroundMarkersAmount; index++) {
+    GEOJSON.features.filter(async (marker: any) => {
+      if (marker.properties.name) {
+        const markerName = JSON.stringify(marker.properties.name).replace(/["​]/g, '');
+  
+        if (markerName.startsWith('sub')) {
+          const sub = Number(markerName.replace(/\D/gim, ''));
+          
+          if (sub === index) {
+            const photosInMarker = JSON.parse(marker.properties.com_exlyo_mapmarker_images_with_ext);
+
+            await photosInMarker.forEach((photo: any, i: number) => {
+              // 0 is even
+              if (i % 2 === 0) undergroundsInLeftColumn.push(photo);
+              if (i % 2 !== 0) undergroundsInRightColumn.push(photo);
+              
+              if (undergroundsInLeftColumn.length + undergroundsInRightColumn.length === undergroundPhotosAmount) {
+                createUndergroundColumns();
+              }
+            })
+          }
+        }
+      }
+    })
+  }
+}
+
+function createUndergroundColumns() {
+  let leftPhotoName = 1,
+      rightPhotoName = 2;
+
+  undergroundsInLeftColumn.forEach(async (leftPhoto, i) => {
+    let rightPhotoRef = undergroundsInRightColumn[i],
+        rightPhoto;
+    
+    if (rightPhotoRef) {
+      rightPhoto = `${kmlMediaDir}/${rightPhotoRef.file_rel_path}${rightPhotoRef.file_extension}`;
+    } else {
+      rightPhoto = noPhoto;
+    }
+
+    let  pageBreak;
+    if (photos.length % 4 === 0 && photos.length !== 0) pageBreak = 'before';
+    
+    const leftAndRightColumnWithTablesAndPhotos = {
+      style: 'columns',
+      columns: [
+        {
+          width: '50%',
+          table: {
+            widths: [240],
+            heights: [12, photoHeight],
+            body: [
+              [
+                {
+                  style: 'titlePhotoTable',
+                  text: `Subterrâneo ${leftPhotoName}`,
+                }
+              ],
+              [
+                {
+                  image: `${kmlMediaDir}/${leftPhoto.file_rel_path}${leftPhoto.file_extension}`,
+                  width: photoWidth,
+                  height: photoHeight,
+                }
+              ],
+            ]
+          },
+        },
+        {
+          width: '50%',
+          table: {
+            widths: [240],
+            heights: [12, photoHeight],
+            body: [
+              [
+                {
+                  style: 'titlePhotoTable',
+                  text: `Subterrâneo ${rightPhotoName}`
+                }
+              ],
+              [
+                {
+                  image: rightPhoto,
+                  width: photoWidth,
+                  height: photoHeight,
+                }
+              ],
+            ]
+          },
+        },
+      ],
+      columnGap: photoGap,
+      pageBreak,
+    };
+
+    photos.push(leftAndRightColumnWithTablesAndPhotos);
+
+    leftPhotoName+=2;
+    rightPhotoName+=2;
+    
+    const subsInMemory = undergroundsInLeftColumn.length + undergroundsInRightColumn.length;
+    
+    if (photos.length === Math.ceil(subsInMemory / 2)) {
+      createPDF();
+      setTimeout(() => {
+        mergePDFS();
+      }, 1000)
     }
   })
 }
@@ -366,12 +486,13 @@ function mergePDFS() {
     fs.rmSync(`${pdfsToMergeDir}/${pdf}`);
   }
 
-  merger.save(`output/relatorio.pdf`);
+  merger.save('output/relatorio.pdf');
+  spinner.stop(true);
 
-  console.log(`${chalk.greenBright(`RELATÓRIO GERADO`)} ${chalk.gray(`[${(Date.now() - startTime) / 1000}s]`)}`);
   console.log('');
-  console.log(`Total de postes: ${chalk.rgb(255, 110, 0)(polesAmountImmutable)}`);
-  console.log(`Fotos no kmz: ${chalk.rgb(255, 110, 0)(photosAmount)}`);
-  console.log(`Postes sem foto: ${chalk.rgb(255, 110, 0)(polesWithoutPhoto)}`);
-  // console.log(`Marcadores ignorados: ${chalk.rgb(255, 110, 0)(Math.floor(ignoredMarkers / POLES_IN_PDF))}`);
+  console.log(boxen(
+    chalk.italic(`Postes: ${polesAmountImmutable}\nSubterrâneos: ${undergroundPhotosAmount}\nFotos no kmz: ${photosAmount}\nSem foto: ${polesWithoutPhoto}`),
+    { title: `${chalk.greenBright(`RELATÓRIO GERADO`)} ${chalk.gray(`[${(Date.now() - startTime) / 1000}s]`)}`,
+    titleAlignment: 'center', textAlignment: 'center', borderColor: 'greenBright', padding: 2 })
+  );
 }
